requirements.txt
Django~=5.0.1
pip~=23.2.1
djangorestframework==3.14.0
wheel~=0.41.2
asgiref~=3.7.2
sqlparse~=0.4.4
typing_extensions~=4.9.0
setuptools~=68.2.0
